PIM-4416: move attribute comparators in a specific directory
