+++ src/Pim/Component/Catalog/Comparator/Attribute/ScalarComparator.php
@@ -1,15 +1,17 @@
 <?php
 
-namespace Pim\Component\Catalog\Comparator;
+namespace Pim\Component\Catalog\Comparator\Attribute;
+
+use Pim\Component\Catalog\Comparator\ComparatorInterface;
 
 /**
- * Comparator which calculate change set for booleans
+ * Comparator which calculate change set for scalars
  *
- * @author    JM Leroux <jean-marie.leroux@akeneo.com>
+ * @author    Gildas Quemener <gildas@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class BooleanComparator implements ComparatorInterface
+class ScalarComparator implements ComparatorInterface
 {
     /**
      * {@inheritdoc}
@@ -17,7 +19,11 @@ class BooleanComparator implements ComparatorInterface
     public function supports($type)
     {
         return in_array($type, [
-            'pim_catalog_boolean',
+            'pim_catalog_date',
+            'pim_catalog_identifier',
+            'pim_catalog_number',
+            'pim_catalog_text',
+            'pim_catalog_textarea'
         ]);
     }
 
@@ -29,10 +35,6 @@ class BooleanComparator implements ComparatorInterface
         $default = ['locale' => null, 'scope' => null, 'value' => null];
         $originals = array_merge($default, $originals);
 
-        if ((bool) $originals['value'] === (bool) $data['value']) {
-            return null;
-        }
-
-        return $data;
+        return $data['value'] !== $originals['value'] ? $data : null;
     }
 }
